+++ examples/brass.rs
@@ -1,5 +1,5 @@
-use fon::{mono::Mono64, Audio};
-use twang::{Mix, Pink, Synth};
+use fon::{mono::Mono64, Audio, Sink};
+use twang::{Mix, Pink, Synth, Signal, Fc};
 
 mod wav;
 
@@ -7,16 +7,9 @@ mod wav;
 const S_RATE: u32 = 48_000;
 
 fn main() {
-    // Initialize audio with five seconds of silence.
-    let mut audio = Audio::<Mono64>::with_silence(S_RATE, S_RATE as usize * 5);
-    // Create the synthesizer.
-    let mut synth = Synth::new();
-    // Create the pink noise generator.
-    let mut pink = Pink::new();
-    // Generate audio samples.
-    synth.gen(audio.sink(..), |fc| {
+    fn brass(pink: &mut Pink, fc: Fc) -> Signal {
         let pink = pink.noise();
-        let tone = fc.freq(220.0).gain(12.0).clamp().gain(0.75);
+        let tone = fc.freq(220.0).gain(12.0).clamp().gain(0.1);
         let airy = tone.abs().gain(pink.abs());
 
         let pone = fc.freq(220.0).gain(12.0).clamp().abs();
@@ -24,7 +17,14 @@ fn main() {
         let main = pone.gain(ptwo);
 
         [airy, main].mix()
-    });
+    }
+
+    // Initialize audio with five seconds of silence.
+    let mut audio = Audio::<Mono64>::with_silence(S_RATE, S_RATE as usize * 5);
+    // Create the synthesizer.
+    let mut synth = Synth::new(Pink::new(), brass);
+    // Generate audio samples.
+    audio.sink(..).stream(&mut synth);
 
     // Write synthesized audio to WAV file.
     wav::write(audio, "brass.wav").expect("Failed to write WAV file");