+++ examples/pink.rs
@@ -1,5 +1,5 @@
-use fon::{mono::Mono64, Audio};
-use twang::{Pink, Synth};
+use fon::{mono::Mono64, Audio, Sink};
+use twang::{Pink, Synth, Fc, Signal};
 
 mod wav;
 
@@ -7,14 +7,17 @@ mod wav;
 const S_RATE: u32 = 48_000;
 
 fn main() {
+    fn gen_pink(pink: &mut Pink, _fc: Fc) -> Signal {
+        pink.noise()
+    }
+
     // Initialize audio with five seconds of silence.
     let mut audio = Audio::<Mono64>::with_silence(S_RATE, S_RATE as usize * 5);
     // Create the synthesizer.
-    let mut synth = Synth::new();
-    // Create the white noise generator.
-    let mut pink = Pink::new();
+    let mut synth = Synth::new(Pink::new(), gen_pink);
+
     // Generate audio samples.
-    synth.gen(audio.sink(..), |_fc| pink.noise());
+    audio.sink(..).stream(&mut synth);
 
     // Write synthesized audio to WAV file.
     wav::write(audio, "pink.wav").expect("Failed to write WAV file");