+++ examples/synth.rs
@@ -17,8 +17,8 @@
 //! let carrier = fc.freq(440.0).shift(message).sine()
 //! ```
 
-use fon::{mono::Mono64, Audio};
-use twang::Synth;
+use fon::{mono::Mono64, Audio, Sink};
+use twang::{Synth, Fc, Signal};
 
 mod wav;
 
@@ -26,16 +26,19 @@ mod wav;
 const S_RATE: u32 = 48_000;
 
 fn main() {
-    // Initialize audio with five seconds of silence.
-    let mut audio = Audio::<Mono64>::with_silence(S_RATE, S_RATE as usize * 5);
-    // Create the synthesizer.
-    let mut synth = Synth::new();
-    // Generate audio samples.
-    synth.gen(audio.sink(..), |fc| {
+    fn gen_synth(_: &mut (), fc: Fc) -> Signal {
         let freq_modulator: f64 = fc.freq(880.0).sine().into();
         let norm_modulator = (freq_modulator + 1.0) * 0.5;
         fc.freq(220.0 * norm_modulator).sine().gain(0.7)
-    });
+    }
+
+    // Initialize audio with five seconds of silence.
+    let mut audio = Audio::<Mono64>::with_silence(S_RATE, S_RATE as usize * 5);
+    // Create the synthesizer.
+    let mut synth = Synth::new((), gen_synth);
+
+    // Generate audio samples.
+    audio.sink(..).stream(&mut synth);
 
     // Write synthesized audio to WAV file.
     wav::write(audio, "synth.wav").expect("Failed to write WAV file");