+++ examples/triangle.rs
@@ -1,5 +1,5 @@
-use fon::{mono::Mono64, Audio};
-use twang::Synth;
+use fon::{mono::Mono64, Audio, Sink};
+use twang::{Synth, Fc, Signal};
 
 mod wav;
 
@@ -7,12 +7,17 @@ mod wav;
 const S_RATE: u32 = 48_000;
 
 fn main() {
+    fn triangle(_: &mut (), fc: Fc) -> Signal {
+        fc.freq(220.0).triangle().gain(0.7)
+    }
+
     // Initialize audio with five seconds of silence.
     let mut audio = Audio::<Mono64>::with_silence(S_RATE, S_RATE as usize * 5);
     // Create the synthesizer.
-    let mut synth = Synth::new();
+    let mut synth = Synth::new((), triangle);
+
     // Generate audio samples.
-    synth.gen(audio.sink(..), |fc| fc.freq(220.0).triangle().gain(0.7));
+    audio.sink(..).stream(&mut synth);
 
     // Write synthesized audio to WAV file.
     wav::write(audio, "triangle.wav").expect("Failed to write WAV file");