+++ examples/voice.rs
@@ -1,5 +1,5 @@
-use fon::{mono::Mono64, Audio};
-use twang::Synth;
+use fon::{mono::Mono64, Audio, Sink};
+use twang::{Synth, Fc, Signal};
 
 mod wav;
 
@@ -7,14 +7,17 @@ mod wav;
 const S_RATE: u32 = 48_000;
 
 fn main() {
+    fn voice(_: &mut (), fc: Fc) -> Signal {
+        fc.freq(440.0).abs().gain(fc.freq(440.0).sine())
+    }
+
     // Initialize audio with five seconds of silence.
     let mut audio = Audio::<Mono64>::with_silence(S_RATE, S_RATE as usize * 5);
     // Create the synthesizer.
-    let mut synth = Synth::new();
+    let mut synth = Synth::new((), voice);
+
     // Generate audio samples.
-    synth.gen(audio.sink(..), |fc| {
-        fc.freq(440.0).abs().gain(fc.freq(440.0).sine())
-    });
+    audio.sink(..).stream(&mut synth);
 
     // Write chord to file
     wav::write(audio, "voice.wav").expect("Failed to write WAV file");