+++ examples/white.rs
@@ -1,5 +1,5 @@
-use fon::{mono::Mono64, Audio};
-use twang::{Synth, White};
+use fon::{mono::Mono64, Audio, Sink};
+use twang::{Synth, White, Fc, Signal};
 
 mod wav;
 
@@ -7,14 +7,17 @@ mod wav;
 const S_RATE: u32 = 48_000;
 
 fn main() {
+    fn gen_white(white: &mut White, _fc: Fc) -> Signal {
+        white.noise()
+    }
+
     // Initialize audio with five seconds of silence.
     let mut audio = Audio::<Mono64>::with_silence(S_RATE, S_RATE as usize * 5);
     // Create the synthesizer.
-    let mut synth = Synth::new();
-    // Create the white noise generator.
-    let mut white = White::new();
+    let mut synth = Synth::new(White::new(), gen_white);
+
     // Generate audio samples.
-    synth.gen(audio.sink(..), |_fc| white.noise());
+    audio.sink(..).stream(&mut synth);
 
     // Write synthesized audio to WAV file.
     wav::write(audio, "white.wav").expect("Failed to write WAV file");